Hotfix parser, atoi changes and more...
- Major changement in parser, handling now a quote argument  with all numbers to sort separated by spaces. Remains to handle argument sent one by one.
- atoi now returns a long type value not int anymore.
- A custom struct with an integer tab and its size will be our stack.

+++ push_swap/libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: cglandus <cglandus@student.42angoulem      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 19:08:03 by cglandus          #+#    #+#             */
-/*   Updated: 2022/11/14 07:26:53 by cglandus         ###   ########.fr       */
+/*   Updated: 2023/12/06 22:33:44 by cglandus         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int				i;
 	int				sign;
